requirements.txt
langchain 
langchain-community 
langchain-core 
transformers 
langchain-google-genai 
google-generativeai
jq 
fastapi
duckduckgo-search
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
re 
json  
langchain_groq
matplotlib
numpy
pandas
seaborn
yfinance
plotly
kaleido
astunparse
httpx
crewai